+++ nuklear/pagehandlers/receive.go
@@ -1,4 +1,4 @@
-package handlers
+package pagehandlers
 
 import (
 	"image"
