+++ nuklear/pagehandlers/sync.go
@@ -1,4 +1,4 @@
-package handlers
+package pagehandlers
 
 import (
 	"fmt"
